requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.2
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
twilio==9.4.1
urllib3==2.2.3
yarl==1.18.3
